overtalkerr/requirements.txt
# Web Framework
Flask==3.1.0
Werkzeug==3.1.3

# Voice Assistant SDKs
ask-sdk-core==1.19.0
ask-sdk-model==1.45.0
ask-sdk-webservice-support==1.19.0  # For webhook verification

# HTTP Client
requests==2.32.3

# Configuration
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.36
alembic==1.14.0

# Production Server
gunicorn==23.0.0

# Logging
python-json-logger==3.2.1

# Utilities
pytz==2024.2

# Enhanced Search & NLP
rapidfuzz==3.10.1  # Fast fuzzy string matching
dateparser==1.2.0  # Natural language date parsing
